requirements.txt
fastapi
pydantic
uvicorn
ppsycopg2
pydantic-sqlalchemy
sqlalchemy-utils
passlib
bcrypt
python-jose
python-multipart